search and fix template

+++ resources/views/shopping/index.blade.php
@@ -58,7 +58,7 @@
                 },
                 success: function (data) {
                     console.log(data);
-                        document.location.href = '/';
+                    location.reload();
                 },
                 error: function() {
                     console.log("Có lỗi xảy ra, vui lòng thử lại.");
@@ -67,26 +67,29 @@
         }
 
         function deleteALl() {
-            var allVals = [];
-            $('.checkboxItem:checked').each(function() {
-                allVals.push($(this).val());
-            });
-            $.ajax({
-                url: "{{ route('shopping.deleteAll') }}",
-                method: 'POST',
-                cache: false,
-                data: {
-                    _token: "{{ csrf_token() }}",
-                    product_id: allVals,
-                },
-                success: function (data) {
-                    console.log(data);
-                        document.location.href = '/';
-                },
-                error: function() {
-                    console.log("Có lỗi xảy ra, vui lòng thử lại.");
-                }
-            });
+            var confirmDel = confirm("Bạn chắc chắn muốn xóa đã chọn không?");
+            if(confirmDel == true) {
+                var allVals = [];
+                $('.checkboxItem:checked').each(function() {
+                    allVals.push($(this).val());
+                });
+                $.ajax({
+                    url: "{{ route('shopping.deleteAll') }}",
+                    method: 'POST',
+                    cache: false,
+                    data: {
+                        _token: "{{ csrf_token() }}",
+                        product_id: allVals,
+                    },
+                    success: function (data) {
+                        console.log(data);
+                        location.reload();
+                    },
+                    error: function() {
+                        console.log("Có lỗi xảy ra, vui lòng thử lại.");
+                    }
+                });
+            }
         }
     </script>
 @endpush
@@ -101,17 +104,18 @@
     {{session('er')}}
     <div class="nav justify-content-center py-1 mb-4">
         <nav class="nav d-flex justify-content-between">
-            <a class="btn btn-light" href="{{ route('category_index')}}">+</a>
-            <a class="btn btn-light ml-2 active" href="">Tất cả</a>
+            <a class="btn btn-light mb-2" href="{{ route('category_index')}}">+</a>
+            <a class="btn btn-light ml-2 mb-2 active" href="">Tất cả</a>
             @foreach($resul_category as $cate)
-                <a class="btn btn-light ml-2" href="{{ url('category')}}/{{$cate->id}}">{{$cate->name_cat}}</a>
+                <a class="btn btn-light ml-2 mb-2" href="{{ url('category')}}/{{$cate->id}}">{{$cate->name_cat}}</a>
             @endforeach
         </nav>
     </div>
             
-    <form class="row justify-content-center">
+    <form class="row justify-content-center" action="{{ route('shopping.searchProduct') }}" method="GET">
+        @csrf
         <div class="col-4">
-          <input type="search" class="form-control" id="search" placeholder="Nhập ..." >
+          <input type="search" name="search" class="form-control" id="search" placeholder="Nhập ..." >
         </div>
         <div class="">
           <button type="submit" class="btn btn-secondary mb-3">Tìm kiếm</button>
@@ -127,7 +131,7 @@
         <a href="{{ route('shopping.create') }}" class="btn btn-info">Thêm Sản Phẩm</a>
     </div>
     
-    <div class="table">
+    <div class="table-responsive mb-3">
         <table class="table">
             <tr>
                 <th class="col-sm-1"><input type="checkbox" id="checkall" class="checkboxAll" name="item[]"></th>
@@ -140,13 +144,13 @@
             @forelse($resul_product as $sp)
             <tr>
                 <td class="col-sm-1"><input type="checkbox" name="item[]" class="checkboxItem" value="{{$sp->id}}"></td>
-                <td class="col-sm-2"><img src="uploads/{{$sp->image}}" height="50" width="100"></td>
+                <td class="col-sm-2"><img src="uploads/{{$sp->image}}" style="object-fit: cover;" height="50" width="80"></td>
                 <td class="col-sm-2">{{$sp->title}}</td>
                 <td class="col-sm-3">{{$sp->content}}</td>
                 <td class="col-sm-2">{{$sp->category->name_cat}}</td>
                 <td class="col-sm-2">
-                    <a href="{{ url('edit')}}/{{$sp->id}}" class="btn btn-primary">Chỉnh sửa</a>
-                    <a href="{{ url('destroy')}}/{{$sp->id}}" class="btn btn-danger">Xóa</a>
+                    <a href="{{ url('edit')}}/{{$sp->id}}" class="btn btn-primary" style="margin: 0 4px 4px 0;">Chỉnh sửa</a>
+                    <a href="{{ url('destroy')}}/{{$sp->id}}" class="btn btn-danger" onclick="return confirm('Bạn chắc chắn muốn xóa không?')" style="margin: 0 4px 4px 0;">Xóa</a>
                 </td>
                 @empty
                 <td colspan="6" style="color: red;">Không có sản phẩm nào</td>
@@ -163,26 +167,30 @@
         </div>
     </div>
 
-    <div class="table">
-        <table class="table">
-            <tr>
-                <th colspan="6"> Tổng sản phẩm đã chọn ({{$count}})</th>
-            </tr>
+    <div class="table-responsive">
+        <table class="table" style="background: #f3f3f375;">
+            <thead>
+                <tr>
+                    <th colspan="6"> Tổng sản phẩm đã chọn ({{$count}})</th>
+                </tr>
+            </thead>
+            <tbody>
             @forelse($resul_product_complete as $sp)
                 <tr>
-                    <td class="col-sm-1"><input type="checkbox" id="item1" name="item[]" readonly="false" value="{{$sp->id}}" checked disabled="true"></td>
-                    <td class="col-sm-2"><img src="uploads/{{$sp->image}}" height="50" width="100"></td>
+                    <td class="col-sm-1"><input type="checkbox" checked disabled="true"></td>
+                    <td class="col-sm-2"><img src="uploads/{{$sp->image}}" style="object-fit: cover;" height="50" width="80"></td>
                     <td class="col-sm-2" style="text-decoration: line-through; color: #80868b!important;">{{$sp->title}}</td>
                     <td class="col-sm-3" style="text-decoration: line-through; color: #80868b!important;">{{$sp->content}}</td>
                     <td class="col-sm-2" style="text-decoration: line-through; color: #80868b!important;">{{$sp->category->name_cat}}</td>
-                    <td class="col-sm-2" width="200px">
-                        <a href="{{ url('category_uncomplete')}}/{{$sp->id}} " class="btn btn-primary">Hoàn Tác</a>
-                        <a href="{{ url('destroy')}}/{{$sp->id}}" class="btn btn-danger">Xóa</a>
+                    <td class="col-sm-2">
+                        <a href="{{ url('category_uncomplete')}}/{{$sp->id}} " class="btn btn-primary" style="margin: 0 4px 4px 0;">Hoàn Tác</a>
+                        <a href="{{ url('destroy')}}/{{$sp->id}}" class="btn btn-danger" onclick="return confirm('Bạn chắc chắn muốn xóa không?')" style="margin: 0 4px 4px 0;">Xóa</a>
                     </td>
                     @empty
                     <td colspan="6" style="color: red;">Chưa có sản phẩm nào được tìm thấy hoặc mua</td>
                 </tr>
             @endforelse
+            </tbody>
         </table>
     </div>
 @stop
